cast message size as uint to fix very large messages (fixes #1591)

+++ com.unity.netcode.gameobjects/Runtime/Messaging/MessagingSystem.cs
@@ -351,7 +351,7 @@ namespace Unity.Netcode
 
             var header = new MessageHeader
             {
-                MessageSize = (ushort)tmpSerializer.Length,
+                MessageSize = (uint)tmpSerializer.Length,
                 MessageType = m_MessageTypes[typeof(TMessageType)],
             };
             BytePacker.WriteValueBitPacked(headerSerializer, header.MessageType);
